Rename dsp_data to dsp_graph
